Moved Method to review/update AppTokenActions under AppTokenActions folder instead
The new method handles authorization better and can review individual tokenActions instead of every action of a token. This is needed since a token can have multiple services in it and 1 person might not be approver of every action.
The command also updates Token.State to Reviewed if every tokenAction.state != Pending.