RequestServiceActions method with new Dto
Updated old CreateAppTokenActions method to be a update method instead. It now handles the case off a user requesting Service 1 that has 3 actions.
Firstly the user request 2 out 3 actions and they are created in the AppTokenActions Dbset. Then he edits his request and only wants 1 out of 3. The method should remove 1 and not add the duplicate again.